backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
asyncpg==0.29.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Browser automation removed for simplified deployment

# AI and ML
openai==1.3.7
google-generativeai==0.3.2
google-cloud-aiplatform==1.38.1
google-cloud-vision==3.4.4
Pillow==10.1.0

# Background tasks
celery==5.3.4

# Cloud services
google-cloud-storage==2.10.0
google-cloud-scheduler==2.12.0
google-cloud-firestore==2.13.1

# Notifications
twilio==8.10.0
sendgrid==6.10.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
python-dateutil==2.8.2
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
aiofiles
black==23.11.0
isort==5.12.0
flake8==6.1.0 

# Web scraping and HTTP
aiohttp>=3.8.0
beautifulsoup4>=4.11.0 